Tell magit about ~/hackery/ dir
Now if I'm in some buffer that isn't under version control, running
magit-status will present a list of my repos under ~/hackery/ so I can
choose which to look at. This is instead of having to then open some
file that is in the repo I want and then run magit-status.